dashboard/requirements.txt
streamlit
plotly
pandas
numpy
scikit-learn
psycopg2-binary
google-api-python-client